moved inventory filters into their own package within the inventory package
