+++ lib/src/main/java/com/group3/racingbot/inventory/filter/SkillFilterable.java
@@ -1,4 +1,4 @@
-package com.group3.racingbot.inventory;
+package com.group3.racingbot.inventory.filter;
 
 /**
  * Classes which implement this can be filtered by a concrete InventoryIteratorDecorator which filters by some skill related property.
